Moved ShedulerProperties class into configuration package
