+++ spring-boot-email-core/src/test/java/it/ozimov/springboot/mail/BaseRedisTest.java
@@ -19,7 +19,7 @@ package it.ozimov.springboot.mail;
 import com.google.common.collect.ImmutableSet;
 import it.ozimov.springboot.mail.configuration.EmailEmbeddedRedis;
 import it.ozimov.springboot.mail.model.EmailSchedulingData;
-import it.ozimov.springboot.mail.service.defaultimpl.SchedulerProperties;
+import it.ozimov.springboot.mail.configuration.SchedulerProperties;
 import lombok.NonNull;
 import org.junit.Before;
 import org.springframework.beans.factory.annotation.Autowired;
