+++ spring-boot-email-core/src/test/java/it/ozimov/springboot/mail/service/defaultimpl/SchedulerPropertiesContextBasedTest.java
@@ -17,6 +17,7 @@
 package it.ozimov.springboot.mail.service.defaultimpl;
 
 import it.ozimov.springboot.mail.ContextBasedTest;
+import it.ozimov.springboot.mail.configuration.SchedulerProperties;
 import org.assertj.core.api.JUnitSoftAssertions;
 import org.junit.Rule;
 import org.junit.Test;
